H operation and computation parallelization
The H operation has been introduced. Each computation has its own
coefficient. When an hadamard gate is applied to a bit, the coefficient
is multiplied by 1 over sqrt(2) and the state is saved, then the bit
(and also the sign depending on the case) are flipped and the
computation is continued. At the end of the current computation, the
saved one gets loaded and executed. With this method for each H gate,
the computation bifurcates in two with smaller coefficients.